refactors props to simplify whats in ChartBlock
